# DOCS/openapi.yaml
openapi: 3.0.3
info:
  title: API do Sistema de Agendamento — MVP
  version: 1.0.0
  description: Contrato do MVP (Fase 1). Datas em UTC (ISO 8601); valores monetários em BRL. Erros seguem RFC 7807
    com o campo `codigo`. O estabelecimento é sempre derivado do token (rotas públicas usam o cabeçalho `X-Estabelecimento`).
servers:
- url: https://localhost:7080/api/v1
  description: Desenvolvimento
- url: https://{host}/api/v1
  description: Produção
  variables:
    host:
      default: api.exemplo.com
...
paths:
...
  /agenda/horarios-livres:
    get:
      tags:
      - Agenda
      operationId: Agenda_HorariosLivres
      summary: Horários livres para os serviços escolhidos
      description: '**Perfis:** Admin, Recepcao, Profissional, Cliente.


        **Erros:** 400 VAL_INVALIDO; 404 RECURSO_NAO_ENCONTRADO.'
      x-perfis:
      - Admin
      - Recepcao
      - Profissional
      - Cliente
      security:
      - bearerAuth: []
      parameters:
      - name: unidadeId
        in: query
        required: true
        schema:
          type: integer
          format: int64
      - name: servicoIds
        in: query
        required: true
        schema:
          type: array
          items:
            type: integer
            format: int64
        style: form
        explode: false
        description: 'Lista separada por vírgula (ex.: 1,2)'
      - name: data
        in: query
        required: true
        schema:
          type: string
          format: date
      - name: profissionalId
        in: query
        required: false
        schema:
          type: integer
          format: int64
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HorariosLivresResponse'
        '400':
          $ref: '#/components/responses/E400'
        '401':
          $ref: '#/components/responses/E401'
        '404':
          $ref: '#/components/responses/E404'
        '500':
          $ref: '#/components/responses/E500'
...
  /agendamentos:
    post:
      tags:
      - Agendamentos
      operationId: Agendamentos_Criar
      summary: Cria agendamento (status PENDENTE)
      description: '**Perfis:** Admin, Recepcao, Cliente.


        **Erros:** 400 VAL_INVALIDO; 404 RECURSO_NAO_ENCONTRADO; 409 AGE_HORARIO_OCUPADO | AGE_BLOQUEIO | AGE_CLIENTE_CONFLITO
        | IDEMPOTENCIA_CONFLITO; 422 AGE_FORA_DISPONIBILIDADE | AGE_PROFISSIONAL_NAO_HABILITADO | AGE_ANTECEDENCIA_INVALIDA.'
      x-perfis:
      - Admin
      - Recepcao
      - Cliente
      security:
      - bearerAuth: []
      parameters:
      - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CriarAgendamentoRequest'
      responses:
        '201':
          description: Criado
          headers:
            Location:
              description: URL do recurso criado
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Agendamento'
        '400':
          $ref: '#/components/responses/E400'
        '401':
          $ref: '#/components/responses/E401'
        '403':
          $ref: '#/components/responses/E403'
        '404':
          $ref: '#/components/responses/E404'
        '409':
          $ref: '#/components/responses/E409'
        '422':
          $ref: '#/components/responses/E422'
        '500':
          $ref: '#/components/responses/E500'
    get:
      tags:
      - Agendamentos
      operationId: Agendamentos_Listar
      summary: Lista agendamentos
      description: '**Perfis:** Admin, Recepcao, Profissional (somente o próprio), Cliente (somente o próprio).


        **Erros:** 400 VAL_INVALIDO.'
      x-perfis:
      - Admin
      - Recepcao
      - Profissional
      - Cliente
      security:
      - bearerAuth: []
      parameters:
      - name: unidadeId
        in: query
        required: false
        schema:
          type: integer
          format: int64
      - name: profissionalId
        in: query
        required: false
        schema:
          type: integer
          format: int64
      - name: clienteId
        in: query
        required: false
        schema:
          type: integer
          format: int64
      - name: status
        in: query
        required: false
        schema:
          $ref: '#/components/schemas/StatusAgendamento'
      - name: de
        in: query
        required: false
        schema:
          type: string
          format: date-time
      - name: ate
        in: query
        required: false
        schema:
          type: string
          format: date-time
      - $ref: '#/components/parameters/Pagina'
      - $ref: '#/components/parameters/Tamanho'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginaAgendamentoResumo'
        '400':
          $ref: '#/components/responses/E400'
        '401':
          $ref: '#/components/responses/E401'
        '403':
          $ref: '#/components/responses/E403'
        '500':
          $ref: '#/components/responses/E500'
...
  /agendamentos/{id}/cancelar:
    post:
      tags:
      - Agendamentos
      operationId: Agendamentos_Cancelar
      summary: Cancela; aplica multa se fora do prazo
      description: '**Perfis:** Admin, Recepcao, Cliente (somente o próprio).


        **Erros:** 404 RECURSO_NAO_ENCONTRADO; 409 AGE_TRANSICAO_INVALIDA.'
      x-perfis:
      - Admin
      - Recepcao
      - Cliente
      security:
      - bearerAuth: []
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CancelarRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CancelarResponse'
        '401':
          $ref: '#/components/responses/E401'
        '403':
          $ref: '#/components/responses/E403'
        '404':
          $ref: '#/components/responses/E404'
        '409':
          $ref: '#/components/responses/E409'
        '500':
          $ref: '#/components/responses/E500'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Pagina:
      name: pagina
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    Tamanho:
      name: tamanho
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      description: GUID para repetir a chamada sem duplicar efeitos
      schema:
        type: string
        format: uuid
  responses:
    E400:
      description: Requisição inválida (VAL_INVALIDO)
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    E401:
      description: Não autenticado
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    E403:
      description: Sem permissão (AUTH_PERMISSAO)
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    E404:
      description: Não encontrado (RECURSO_NAO_ENCONTRADO)
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    E409:
      description: Conflito de estado
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    E422:
      description: Regra de negócio violada
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
...
    E500:
      description: Erro interno
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
...
  schemas:
...
    StatusAgendamento:
      type: string
      enum:
      - PENDENTE
      - CONFIRMADO
      - EM_ATENDIMENTO
      - CONCLUIDO
      - CANCELADO
      - NO_SHOW
      - REMARCADO
...
    OrigemAgendamento:
      type: string
      enum:
      - APP
      - SITE
      - BALCAO
      - TELEFONE
      - WHATSAPP
...
    ProblemDetails:
      type: object
      properties:
        type:
          type: string
        title:
          type: string
        status:
          type: integer
          format: int32
        codigo:
          type: string
          description: Código de erro de negócio (ex. AGE_HORARIO_OCUPADO)
        detail:
          type: string
        instance:
          type: string
        correlationId:
          type: string
        erros:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
          description: Erros por campo (somente 400)
...
    HorarioLivre:
      type: object
      properties:
        profissionalId:
          type: integer
          format: int64
        inicio:
          type: string
          format: date-time
        fim:
          type: string
          format: date-time
    HorariosLivresResponse:
      type: object
      properties:
        data:
          type: string
          format: date
        duracaoTotalMin:
          type: integer
          format: int32
        horarios:
          type: array
          items:
            $ref: '#/components/schemas/HorarioLivre'
...
    ClienteResumo:
      type: object
      properties:
        id:
          type: integer
          format: int64
        nome:
          type: string
    ProfissionalResumo:
      type: object
      properties:
        id:
          type: integer
          format: int64
        nome:
          type: string
    ItemAgendamento:
      type: object
      properties:
        servicoId:
          type: integer
          format: int64
        nome:
          type: string
        duracaoMin:
          type: integer
          format: int32
        precoCongelado:
          type: number
          format: double
    Agendamento:
      type: object
      properties:
        id:
          type: integer
          format: int64
        status:
          $ref: '#/components/schemas/StatusAgendamento'
        inicio:
          type: string
          format: date-time
        fim:
          type: string
          format: date-time
        unidadeId:
          type: integer
          format: int64
        cliente:
          $ref: '#/components/schemas/ClienteResumo'
        profissional:
          $ref: '#/components/schemas/ProfissionalResumo'
        itens:
          type: array
          items:
            $ref: '#/components/schemas/ItemAgendamento'
        valorTotal:
          type: number
          format: double
        origem:
          $ref: '#/components/schemas/OrigemAgendamento'
        criadoEm:
          type: string
          format: date-time
    AgendamentoResumo:
      type: object
      properties:
        id:
          type: integer
          format: int64
        status:
          $ref: '#/components/schemas/StatusAgendamento'
        inicio:
          type: string
          format: date-time
        fim:
          type: string
          format: date-time
        clienteNome:
          type: string
        profissionalId:
          type: integer
          format: int64
        profissionalNome:
          type: string
        valorTotal:
          type: number
          format: double
    CriarAgendamentoRequest:
      type: object
      required:
      - unidadeId
      - profissionalId
      - inicio
      - servicoIds
      properties:
        clienteId:
          type: integer
          format: int64
          description: Ignorado para o perfil Cliente (usa o do token)
        unidadeId:
          type: integer
          format: int64
        profissionalId:
          type: integer
          format: int64
        inicio:
          type: string
          format: date-time
        servicoIds:
          type: array
          minItems: 1
          items:
            type: integer
            format: int64
        origem:
          $ref: '#/components/schemas/OrigemAgendamento'
    CancelarRequest:
      type: object
      properties:
        motivo:
          type: string
    CancelarResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        status:
          $ref: '#/components/schemas/StatusAgendamento'
        multaAplicada:
          type: boolean
        valorMulta:
          type: number
          format: double
...
    PaginaAgendamentoResumo:
      type: object
      properties:
        itens:
          type: array
          items:
            $ref: '#/components/schemas/AgendamentoResumo'
        pagina:
          type: integer
        tamanho:
          type: integer
        totalItens:
          type: integer
        totalPaginas:
          type: integer
...